Rename 'logging' to make it easier to use python `logging`
